+++ homeassistant/components/rest/sensor.py
@@ -10,31 +10,28 @@ import voluptuous as vol
 import xmltodict
 
 from homeassistant.components.sensor import (
-    CONF_STATE_CLASS,
     DOMAIN as SENSOR_DOMAIN,
     PLATFORM_SCHEMA,
     SensorDeviceClass,
-    SensorEntity,
 )
 from homeassistant.components.sensor.helpers import async_parse_date_datetime
 from homeassistant.const import (
-    CONF_DEVICE_CLASS,
     CONF_FORCE_UPDATE,
-    CONF_NAME,
     CONF_RESOURCE,
     CONF_RESOURCE_TEMPLATE,
-    CONF_UNIT_OF_MEASUREMENT,
+    CONF_UNIQUE_ID,
     CONF_VALUE_TEMPLATE,
 )
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import PlatformNotReady
 import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
+from homeassistant.helpers.template_entity import TemplateSensor
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
 from . import async_get_config_and_coordinator, create_rest_data_from_config
-from .const import CONF_JSON_ATTRS, CONF_JSON_ATTRS_PATH
-from .entity import RestEntity
+from .const import CONF_JSON_ATTRS, CONF_JSON_ATTRS_PATH, DEFAULT_SENSOR_NAME
+from .entity import BaseRestEntity
 from .schema import RESOURCE_SCHEMA, SENSOR_SCHEMA
 
 _LOGGER = logging.getLogger(__name__)
@@ -70,67 +67,54 @@ async def async_setup_platform(
             raise PlatformNotReady from rest.last_exception
         raise PlatformNotReady
 
-    name = conf.get(CONF_NAME)
-    unit = conf.get(CONF_UNIT_OF_MEASUREMENT)
-    device_class = conf.get(CONF_DEVICE_CLASS)
-    state_class = conf.get(CONF_STATE_CLASS)
-    json_attrs = conf.get(CONF_JSON_ATTRS)
-    json_attrs_path = conf.get(CONF_JSON_ATTRS_PATH)
-    value_template = conf.get(CONF_VALUE_TEMPLATE)
-    force_update = conf.get(CONF_FORCE_UPDATE)
-    resource_template = conf.get(CONF_RESOURCE_TEMPLATE)
-
-    if value_template is not None:
-        value_template.hass = hass
+    unique_id = conf.get(CONF_UNIQUE_ID)
 
     async_add_entities(
         [
             RestSensor(
+                hass,
                 coordinator,
                 rest,
-                name,
-                unit,
-                device_class,
-                state_class,
-                value_template,
-                json_attrs,
-                force_update,
-                resource_template,
-                json_attrs_path,
+                conf,
+                unique_id,
             )
         ],
     )
 
 
-class RestSensor(RestEntity, SensorEntity):
+class RestSensor(BaseRestEntity, TemplateSensor):
     """Implementation of a REST sensor."""
 
     def __init__(
         self,
+        hass,
         coordinator,
         rest,
-        name,
-        unit_of_measurement,
-        device_class,
-        state_class,
-        value_template,
-        json_attrs,
-        force_update,
-        resource_template,
-        json_attrs_path,
+        config,
+        unique_id,
     ):
         """Initialize the REST sensor."""
-        super().__init__(coordinator, rest, name, resource_template, force_update)
+        BaseRestEntity.__init__(
+            self,
+            coordinator,
+            rest,
+            config.get(CONF_RESOURCE_TEMPLATE),
+            config.get(CONF_FORCE_UPDATE),
+        )
+        TemplateSensor.__init__(
+            self,
+            hass,
+            config=config,
+            fallback_name=DEFAULT_SENSOR_NAME,
+            unique_id=unique_id,
+        )
         self._state = None
-        self._unit_of_measurement = unit_of_measurement
-        self._value_template = value_template
-        self._json_attrs = json_attrs
+        self._value_template = config.get(CONF_VALUE_TEMPLATE)
+        if (value_template := self._value_template) is not None:
+            value_template.hass = hass
+        self._json_attrs = config.get(CONF_JSON_ATTRS)
         self._attributes = None
-        self._json_attrs_path = json_attrs_path
-
-        self._attr_native_unit_of_measurement = self._unit_of_measurement
-        self._attr_device_class = device_class
-        self._attr_state_class = state_class
+        self._json_attrs_path = config.get(CONF_JSON_ATTRS_PATH)
 
     @property
     def native_value(self):
