+++ homeassistant/components/template/alarm_control_panel.py
@@ -125,6 +125,8 @@ async def async_setup_platform(
 class AlarmControlPanelTemplate(TemplateEntity, AlarmControlPanelEntity):
     """Representation of a templated Alarm Control Panel."""
 
+    _attr_should_poll = False
+
     def __init__(
         self,
         hass,
