+++ homeassistant/components/template/button.py
@@ -78,6 +78,8 @@ async def async_setup_platform(
 class TemplateButtonEntity(TemplateEntity, ButtonEntity):
     """Representation of a template button."""
 
+    _attr_should_poll = False
+
     def __init__(
         self,
         hass: HomeAssistant,
