+++ homeassistant/components/template/fan.py
@@ -125,6 +125,8 @@ async def async_setup_platform(
 class TemplateFan(TemplateEntity, FanEntity):
     """A template fan component."""
 
+    _attr_should_poll = False
+
     def __init__(
         self,
         hass,
