+++ homeassistant/components/template/vacuum.py
@@ -126,6 +126,8 @@ async def async_setup_platform(
 class TemplateVacuum(TemplateEntity, StateVacuumEntity):
     """A template vacuum component."""
 
+    _attr_should_poll = False
+
     def __init__(
         self,
         hass,
